Changes in draft and R
In draft: added the Gibbs sampler steps for linear regression model
which is a general case for the sparse normal means. In R code,
implemented the Gibbs sampler in a standalone file called
sql.normal.means. Eventually, there should be different implementations
for the sparse normal means and the linear regression problem because
they require different strategies.